Refactor model objects
