+++ src/main/java/net/boomerangplatform/Application.java
@@ -7,8 +7,7 @@ import org.springframework.boot.autoconfigure.security.servlet.SecurityAutoConfi
 @SpringBootApplication(exclude = { SecurityAutoConfiguration.class })
 public class Application {
 
-  public static void main(String[] args) {
-    SpringApplication.run(Application.class, args);
-  }
-
+    public static void main(String[] args) {
+        SpringApplication.run(Application.class, args);
+    }
 }
