--- src/main/java/net/boomerangplatform/client/model/ValidateTokenRequest.java
@@ -1,14 +1,0 @@
-package net.boomerangplatform.client.model;
-
-public class ValidateTokenRequest {
-
-  private String token;
-
-  public String getToken() {
-    return token;
-  }
-
-  public void setToken(String token) {
-    this.token = token;
-  }
-}
