+++ src/main/java/net/boomerangplatform/model/ValidateTokenRequest.java
@@ -1,0 +1,14 @@
+package net.boomerangplatform.model;
+
+public class ValidateTokenRequest {
+
+    private String token;
+
+    public String getToken() {
+        return token;
+    }
+
+    public void setToken(String token) {
+        this.token = token;
+    }
+}
